Move the count_leading_spaces/1 to Text module
