+++ apps/common/test/lexical/math_test.exs
@@ -28,15 +28,4 @@ defmodule Lexical.MathTest do
       assert Math.clamp(high, low, mid) == mid
     end
   end
-
-  property "count_leading_spaces/1" do
-    check all(
-            maybe_spaces <- string([?\t, ?\s]),
-            string_base <- string(:printable),
-            maybe_with_leading_spaces = maybe_spaces <> string_base
-          ) do
-      space_count = byte_size(maybe_spaces)
-      assert Math.count_leading_spaces(maybe_with_leading_spaces) == space_count
-    end
-  end
 end
